+++ plugins/inputs/prometheus_scraper/metric_prometheus.go
@@ -5,7 +5,24 @@ package prometheus_scraper
 
 import (
 	"github.com/prometheus/prometheus/model/textparse"
+	"github.com/prometheus/prometheus/model/value"
+	"math"
 )
+
+type PrometheusMetricBatch []*PrometheusMetric
+
+type PrometheusMetric struct {
+	tags        map[string]string
+	metricName  string
+	metricValue float64
+	metricType  string
+	timeInMS    int64 // Unix time in milli-seconds
+}
+
+func (pm *PrometheusMetric) isValueValid() bool {
+	//treat NaN and +/-Inf values as invalid as emf log doesn't support them
+	return !value.IsStaleNaN(pm.metricValue) && !math.IsNaN(pm.metricValue) && !math.IsInf(pm.metricValue, 0)
+}
 
 func (pm *PrometheusMetric) isCounter() bool {
 	return pm.metricType == string(textparse.MetricTypeCounter)
