+++ plugins/inputs/prometheus_scraper/metrics_metadata.go
@@ -8,19 +8,6 @@ import (
 	"errors"
 	"github.com/prometheus/prometheus/model/textparse"
 	"github.com/prometheus/prometheus/scrape"
-	"strings"
-)
-
-const (
-	prometheusMetricTypeKey = "prom_metric_type"
-	metricsSuffixCount      = "_count"
-	metricsSuffixBucket     = "_bucket"
-	metricsSuffixSum        = "_sum"
-	metricSuffixTotal       = "_total"
-)
-
-var (
-	trimmableSuffixes = []string{metricsSuffixCount, metricsSuffixBucket, metricsSuffixSum, metricSuffixTotal}
 )
 
 type MetadataCache interface {
@@ -60,25 +47,4 @@ func metadataForMetric(metricName string, mc MetadataCache) *scrape.MetricMetada
 		Metric: metricName,
 		Type:   textparse.MetricTypeUnknown,
 	}
-}
-
-// Get the metric name in the TYPE comments for Summary and Histogram
-// e.g # TYPE nginx_ingress_controller_request_duration_seconds histogram
-//     # TYPE nginx_ingress_controller_ingress_upstream_latency_seconds summary
-
-func normalizeMetricName(name string) string {
-	for _, s := range trimmableSuffixes {
-		if strings.HasSuffix(name, s) && name != s {
-			return strings.TrimSuffix(name, s)
-		}
-	}
-	return name
-}
-
-func isInternalMetric(metricName string) bool {
-	//For each endpoint, Prometheus produces a set of internal metrics. See https://prometheus.io/docs/concepts/jobs_instances/
-	if metricName == "up" || strings.HasPrefix(metricName, "scrape_") {
-		return true
-	}
-	return false
 }
