+++ plugins/inputs/prometheus_scraper/util.go
@@ -7,6 +7,18 @@ import (
 	"bytes"
 	"fmt"
 	"sort"
+	"strings"
+)
+
+const (
+	metricsSuffixCount  = "_count"
+	metricsSuffixBucket = "_bucket"
+	metricsSuffixSum    = "_sum"
+	metricSuffixTotal   = "_total"
+)
+
+var (
+	trimmableSuffixes = []string{metricsSuffixCount, metricsSuffixBucket, metricsSuffixSum, metricSuffixTotal}
 )
 
 func getTagsKey(pm *PrometheusMetric) *bytes.Buffer {
@@ -57,4 +69,25 @@ func mergePrometheusMetrics(mm *metricMaterial, pm *PrometheusMetric) *metricMat
 
 	mm.fields[pm.metricName] = pm.metricValue
 	return mm
+}
+
+// Get the metric name in the TYPE comments for Summary and Histogram
+// e.g # TYPE nginx_ingress_controller_request_duration_seconds histogram
+//     # TYPE nginx_ingress_controller_ingress_upstream_latency_seconds summary
+
+func normalizeMetricName(name string) string {
+	for _, s := range trimmableSuffixes {
+		if strings.HasSuffix(name, s) && name != s {
+			return strings.TrimSuffix(name, s)
+		}
+	}
+	return name
+}
+
+func isInternalMetric(metricName string) bool {
+	//For each endpoint, Prometheus produces a set of internal metrics. See https://prometheus.io/docs/concepts/jobs_instances/
+	if metricName == "up" || strings.HasPrefix(metricName, "scrape_") {
+		return true
+	}
+	return false
 }
